Implement conversion of coordinates to M2's angle
While this may sound like a process that is akin to M1's angle
calculation, here it is a bit different. Namely, instead of rotating
the camera pi radians with m2, we can (and probably should) just rotate
it pi radians with m1. Therefore there exist now 2 ways of calculating
the angle, because this function is able to return negative values. It
is important to note, that in order not to introduce bugs in the future,
the angles are to be though of as completely different data, and more
importantly, should be handled differently.