Fix motor bug + partially improve the queue functionality
The most important change here is for the bug that I discovered when
testing the code for more timely-spaced-out points. The problem was that
if the motors were left alone after reaching their goal, they would not
stop and continue moving when they shouldn't. This was previously
undetectable in testing, since I always defined points to be given a
short amount of time after each other to speed up testing.