Correct checking cases in is_horizontal_reached()
The problem wth the previous implementation is that there was a danger
of the camera getting stuck while trying to reach a point very close to
the x axis. The camera would be pretty much facing parallel with the x
axis, making its rotation 0, while the target would be very close to the
x axis, but still in the positive xs and ys, making the rotation 4096.
This would not be caught by the previous implementation, as
abs(4096-0)=4096>ANGLE_ACCEPTABLE_DEVIATION.